Restructure code to be usable as a bad library.
(*in contrast to unusable as a broken library)

Move poll watching function to CLI side (cmd package).
Also add a GetPoll (singular) function to get a Poll response of a specific poll instead of parsing through all returns.
Rename "server" package to "callback" package because of plans for future expansion that could confuse the package names.