API Server
Add an API server for local external tool control.
For example, a basic Python script or game engine can now interact with Twitch through MSC without needing to deal with any of the Twitch complexities.
This is about 80% tested. I'm not going live on my channel to test the remaining parts, but I'm open to bug reports and PRs if this doesn't work and if I don't catch it before.